feat: change the class of the -card-details- component to functional and decompose it
