+++ src/components/card-details/card-details-render.tsx
@@ -1,0 +1,65 @@
+import type React from 'react';
+import CardDetailsLogic from './card-details-logic';
+import type { CardDetailsProps } from './types/card-details.type';
+
+const CardDetails = ({ url }: CardDetailsProps): React.JSX.Element => {
+  const { details } = CardDetailsLogic({ url });
+
+  if (!details) {
+    return <div>No details available</div>;
+  }
+  return (
+    <>
+      <ul>
+        <li className="flex flex-row gap-2">
+          <h3 className="font-mono text-1.5xl">Height:</h3>
+          <p>{details.height}</p>
+        </li>
+        <li className="flex flex-row gap-2">
+          <h3 className="font-mono text-1.5xl">Weight:</h3>
+          <p>{details.weight}</p>
+        </li>
+        <li className="flex flex-row gap-2">
+          <h3 className="font-mono text-1.5xl">Types:</h3>
+          <p>
+            {details.types.map((typeInfo) => typeInfo.type.name).join(', ')}
+          </p>
+        </li>
+        <li className="flex flex-row gap-2">
+          <h3 className="font-mono text-1.5xl">Abilities:</h3>
+          <p>
+            {details.abilities
+              .map((abilityInfo) => abilityInfo.ability.name)
+              .join(', ')}
+          </p>
+        </li>
+        <li className="flex flex-row gap-2">
+          <h3 className="font-mono text-1.5xl">Stats:</h3>
+          <p>
+            {details.stats
+              .map((statInfo) => `${statInfo.stat.name}: ${statInfo.base_stat}`)
+              .join(', ')}
+          </p>
+        </li>
+        <li className="flex flex-row gap-2">
+          <h3 className="font-mono text-1.5xl">Moves:</h3>
+          <p>
+            {details.moves
+              .slice(0, 5)
+              .map((moveInfo) => moveInfo.move.name)
+              .join(', ')}
+          </p>
+        </li>
+      </ul>
+      {details.sprites.front_default && (
+        <img
+          src={details.sprites.front_default}
+          alt="Pokemon"
+          className="justify-self-center w-40 h-40"
+        />
+      )}
+    </>
+  );
+};
+
+export default CardDetails;
