+++ src/tests/CardDetails.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen, waitFor } from '@testing-library/react';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import '@testing-library/jest-dom';
-import CardDetails from '../components/CardDetails';
+import CardDetails from '../components/card-details/card-details-render';
 
 const mockDetails = {
   height: 10,
